reworked delete as modal and started update

+++ admin/source/js/delete.js
@@ -1,4 +1,4 @@
-function delete_(r, preview=false) {
+function delete_(r) {
   if (r == null) {
     r = {};
   }
@@ -20,14 +20,14 @@ function delete_(r, preview=false) {
   $.post("/", request)
   .done(function (data) {
     last_selected_container = request["of"];
+    $('input[name=into], input[name=of]').val(last_selected_container);
     $('input[name=where]').val(r.where);
     $('input[name=limit]').val(r.limit);
     $('input[name=offset]').val(r.offset);
     $('#current_container').text(last_selected_container);
     $('#total_entrys').text(data.total);
-    if (preview == false) {
-      display_message( {content:"Delete: Removed: "+data.hits+" entry(s)", color:"#ccc"} );
-    }
+    $('#delete_modal').modal('hide');
+    display_message( {content:"Delete: Removed: "+data.hits+" entry(s)", color:"#ccc"} );
   })
   .fail(function (data) {
     data = data.responseJSON ? data.responseJSON : {};
@@ -44,25 +44,31 @@ function delete_(r, preview=false) {
 }
 
 function start_delete() {
-  if ($('#delete_modal').is(':visible')) {
-    $('#delete_modal').collapse('hide');
-    return;
-  }
-  $('#modal-space > .collapse').collapse('hide');
-  $('#delete_modal').find('[name=of]').val( last_selected_container );
+  $('#delete_modal').find('[name=into]').val( last_selected_container )
   $('#delete_modal').find('.need_correction').removeClass('need_correction');
-  $('#delete_modal').collapse('show');
+  $('#delete_modal').modal('show');
 }
 
 function modal_delete() {
   let col_modal = $('#delete_modal');
   let r = {};
+  let table_name = col_modal.find('[name=of]').val();
+  if (table_name == "") {
+    col_modal.find('[name=of]').addClass('need_correction').focus();
+    return ;
+  }
 
-  r['of'] = col_modal.find('[name=of]').val();
+  r['of'] = table_name;
   r['where'] = col_modal.find('[name=where]').val();
   r['limit'] = col_modal.find('[name=limit]').val();
   r['offset'] = col_modal.find('[name=offset]').val();
 
-  return delete_(r);
+  let t = false;
+  if (r.where == "" || r.where == null) {
+    t = confirm("WARNING\nYour 'where' field is empty. that will erase all entrys if not hold by\nlimit or offset.\n\nSure you wanna do this?");
+  }
+  if (t) {
+    return delete_(r);
+  }
 }
